+++ src/components/dialog/dialog.ts
@@ -166,7 +166,7 @@ export default class SlDialog extends LitElement {
 
       // Browsers that support el.focus({ preventScroll }) can set initial focus immediately
       if (hasPreventScroll) {
-        const slInitialFocus = this.slInitialFocus.emit();
+        const slInitialFocus = this.slInitialFocus.emit({ cancelable: true });
         if (!slInitialFocus.defaultPrevented) {
           this.panel.focus({ preventScroll: true });
         }
@@ -182,7 +182,7 @@ export default class SlDialog extends LitElement {
       // Browsers that don't support el.focus({ preventScroll }) have to wait for the animation to finish before initial
       // focus to prevent scrolling issues. See: https://caniuse.com/mdn-api_htmlelement_focus_preventscroll_option
       if (!hasPreventScroll) {
-        const slInitialFocus = this.slInitialFocus.emit();
+        const slInitialFocus = this.slInitialFocus.emit({ cancelable: true });
         if (!slInitialFocus.defaultPrevented) {
           this.panel.focus({ preventScroll: true });
         }
@@ -216,7 +216,7 @@ export default class SlDialog extends LitElement {
   }
 
   handleOverlayClick() {
-    const slOverlayDismiss = this.slOverlayDismiss.emit();
+    const slOverlayDismiss = this.slOverlayDismiss.emit({ cancelable: true });
     if (!slOverlayDismiss.defaultPrevented) {
       this.hide();
     }